Brings the order flow from a two-state stub to a full lifecycle: a state machine
that refuses illegal moves, an orchestrated saga over the outbox holding stock
and capturing payment, revenue counted when money is taken rather than when an
order is confirmed, and fulfilment through to DELIVERED.

Also closes the defects the spec work uncovered along the way - overselling,
unclamped discounts, card data in the aggregate and on the wire, a confirm that
could place a second order, and a buyer cancellation that leaked stock and could
be silently overwritten.

Specs: order-lifecycle, payment, stock-reservation, order-fulfilment,
analytics-revenue, order-lifecycle-remaining-work. ADR-0011 records why the saga
is orchestrated rather than choreographed.

What remains is tracked in docs/specs/order-lifecycle-remaining-work.md, in
criticality order. Refunds are first: this merge makes it possible to take money
and there is no path to give it back.

+++ backend/analytics-service/src/main/java/the/chak/ecommerce/analytics/boundary/dto/AnalyticsResponse.java
@@ -1,5 +1,6 @@
 package the.chak.ecommerce.analytics.boundary.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -20,5 +21,5 @@ public class AnalyticsResponse {
     /** Revenue per category, highest first. */
     private List<CategoryRevenue> categoryBreakdown;
 
-    private Double totalRevenue;
+    private BigDecimal totalRevenue;
 }
